+++ scripts/shared.py
@@ -1,0 +1,3 @@
+columns = ['name','prompt','negative_prompt']
+user_columns = ['prompt','negative_prompt','notes']
+display_columns = ['sort', 'name','prompt','negative_prompt','notes']